+++ minimalkv/_hstores.py
@@ -1,9 +1,9 @@
 import os
 
-from minimalkv.contrib import ExtendedKeyspaceMixin
 from minimalkv.fs import FilesystemStore
 from minimalkv.memory import DictStore
 from minimalkv.memory.redisstore import RedisStore
+from minimalkv.mixins import ExtendedKeyspaceMixin
 from minimalkv.net.azurestore import AzureBlockBlobStore
 from minimalkv.net.botostore import BotoStore
 from minimalkv.net.gcstore import GoogleCloudStore