+++ minimalkv/key_value_store.py
@@ -13,8 +13,8 @@ class KeyValueStore:
     Class to access a key-value store.
 
     Supported keys are ascii-strings containing alphanumeric characters or symbols out
-    of ``minimalkv.VALID_NON_NUM`` of length not greater than 250. Values (or records)
-    are stored as raw bytes.
+    of ``minimalkv.constants.VALID_NON_NUM`` of length not greater than 250. Values
+    (or records) are stored as raw bytes.
     """
 
     def __contains__(self, key: str) -> bool: