+++ minimalkv/mixins.py
@@ -1,7 +1,7 @@
 from io import BytesIO
 from typing import IO, Callable, Optional, Union
 
-from minimalkv.constants import FOREVER, NOT_SET
+from minimalkv.constants import FOREVER, NOT_SET, VALID_KEY_RE_EXTENDED
 
 
 class UrlMixin:
@@ -54,18 +54,19 @@ class TimeToLiveMixin:
     """Mixin to allow keys to expire after a certain amount of time.
 
     This mixin overrides some of the signatures of the api of
-    :class:`~minimalkv.KeyValueStore`, albeit in a backwards compatible manner.
+    :class:`~minimalkv.key_value_store.KeyValueStore`, albeit in a backwards compatible
+    manner.
 
     Any value given for a time-to-live parameter must be one of the following:
 
     * A positive ``int`` or ``float``, representing seconds,
-    * ``minimalkv.FOREVER``, meaning no expiration
-    * ``minimalkv.NOT_SET``, meaning that no TTL configuration will be
+    * ``minimalkv.constants.FOREVER``, meaning no expiration
+    * ``minimalkv.constants.NOT_SET``, meaning that no TTL configuration will be
       done at all or
     * ``None`` representing the default (see
       :class:`.TimeToLiveMixin`'s ``default_ttl_secs``).
 
-    .. note:: When deriving from :class:`~minimalkv.TimeToLiveMixin`, the same
+    .. note:: When deriving from :class:`~minimalkv.mixins.TimeToLiveMixin`, the same
        default implementations for ``_put``, ``_put_file`` and
        ``_put_filename`` are provided, except that they all take an additional
        ``ttl_secs`` argument. For more information on how to implement
@@ -92,9 +93,9 @@ class TimeToLiveMixin:
         Parameters
         ----------
         ttl_secs: numeric or string or None
-            Time to live. Numeric or one of ``minimalkv.FOREVER`` or
-            ``minimalkv.NOT_SET``. ``None`` will be replaced with
-            ``minimalkv.TimeToLiveMixin.default_ttl_secs``.
+            Time to live. Numeric or one of ``minimalkv.constants.FOREVER`` or
+            ``minimalkv.constants.NOT_SET``. ``None`` will be replaced with
+            ``minimalkv.mixins.TimeToLiveMixin.default_ttl_secs``.
 
         Raises
         ------
@@ -124,9 +125,9 @@ class TimeToLiveMixin:
         """Store bytestring data at key.
 
         If ``ttl_secs`` is a positive number, the key will expire after ``ttl_secs``.
-        Other possible values for ``ttl_secs`` are ``minimalkv.FOREVER`` (no expiration)
-        and ``minimalkv.NOT_SET`` (no TTL configuration). ``None`` will be replaced with
-        ``minimalkv.TimeToLiveMixin.default_ttl_secs``.
+        Other possible values for ``ttl_secs`` are ``minimalkv.constants.FOREVER`` (no expiration)
+        and ``minimalkv.constants.NOT_SET`` (no TTL configuration). ``None`` will be replaced with
+        ``minimalkv.mixins.TimeToLiveMixin.default_ttl_secs``.
 
         Parameters
         ----------
@@ -172,9 +173,9 @@ class TimeToLiveMixin:
         unnecessary copies. To prevent this, pass the opened file instead.
 
         If ``ttl_secs`` is a positive number, the key will expire after ``ttl_secs``.
-        Other possible values for ``ttl_secs`` are `minimalkv.FOREVER` (no expiration)
-        and ``minimalkv.NOT_SET`` (no TTL configuration). ``None`` will be replaced with
-        ``minimalkv.TimeToLiveMixin.default_ttl_secs``.
+        Other possible values for ``ttl_secs`` are `minimalkv.constants.FOREVER` (no expiration)
+        and ``minimalkv.constants.NOT_SET`` (no TTL configuration). ``None`` will be replaced with
+        ``minimalkv.mixins.TimeToLiveMixin.default_ttl_secs``.
 
         Parameters
         ----------
@@ -381,3 +382,32 @@ class CopyMixin(object):
         self._copy(source, dest)
         self._delete(source)
         return dest
+
+
+class ExtendedKeyspaceMixin:
+    """A mixin to extend the keyspace to allow slashes and spaces in keynames.
+
+    Attention: A single / is NOT allowed.
+    Use it by extending first from ` :class:`~minimalkv.mixins.ExtendedKeyspaceMixin`
+    and then by the desired store.
+    Note: This Mixin is unsupported and might not work correctly with all backends.
+
+    """
+
+    def _check_valid_key(self, key: Optional[str]) -> None:
+        """Check if a key is valid and raises a ValueError if its not.
+
+        When in need of checking a key for validity, always use this
+        method if possible.
+
+        Parameters
+        ----------
+        key : str
+            The key to be checked
+
+        """
+        if key is not None:
+            if not isinstance(key, str):
+                raise ValueError("%r is not a valid key type" % key)
+            elif not VALID_KEY_RE_EXTENDED.match(key) or key == "/":
+                raise ValueError("%r contains illegal characters" % key)