+++ tests/test_mongo.py
@@ -10,8 +10,8 @@ pymongo = pytest.importorskip("pymongo")
 from basic_store import BasicStore
 from conftest import ExtendedKeyspaceTests
 
-from minimalkv.contrib import ExtendedKeyspaceMixin
 from minimalkv.db.mongo import MongoStore
+from minimalkv.mixins import ExtendedKeyspaceMixin
 
 
 class TestMongoDB(BasicStore):